+++ config/defaults.inc.php
@@ -30,7 +30,7 @@ $config = [];
 // Note: for SQLite use absolute path (Linux): 'sqlite:////full/path/to/sqlite.db?mode=0646'
 //       or (Windows): 'sqlite:///C:/full/path/to/sqlite.db'
 // Note: Various drivers support various additional arguments for connection,
-//       for Mysql: key, cipher, cert, capath, ca, verify_server_cert,
+//       for Mysql: key, cipher, cert, capath, ca, verify_server_cert, emulate_prepares
 //       for Postgres: application_name, sslmode, sslcert, sslkey, sslrootcert, sslcrl, sslcompression, service.
 //       e.g. 'mysql://roundcube:@localhost/roundcubemail?verify_server_cert=false'
 $config['db_dsnw'] = 'mysql://roundcube:@localhost/roundcubemail';