+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Add identities management script - bin/identity.sh (#8887)
 - Prefer 8bit over quoted-printable for HTML parts, when force_7bit is disabled (#8477)
 - Convert images in HTML content pasted into HTML editor to `data:` URIs (and later to attachments) (#6938)
+- Add possibility to change ATTR_EMULATE_PREPARES via config file (#9213)
 - ACL: Set default of 'acl_specials' option to ['anyone'] (#8911)
 - Enigma: Support Kolab's Web Of Anti-Trust feature (#8626)
 - Managesieve: Support :encodeurl (RFC 5435) (#8917)